remove c++
