+++ Sources/AdaEngine/Scene/Physics/Physics2D/Components/Physics2DWorldComponent.swift
@@ -6,5 +6,5 @@
 //
 
 public struct Physics2DWorldComponent: Component {
-    public let world: PhysicsWorld2D
+//    public let world: PhysicsWorld2D
 }
